Add FreeBSD ``update scripts''
This commit adds the following FreeBSD-specific Bourne shell scripts:
dbu.sh to download binary patches for FreeBSD base; dports.sh to create
a tarball from the current ports tree; dsrc.sh to create a tarball from
the current source code of the base.  All the scripts are designed to be
executed on a FreeBSD host only.

Signed-off-by: Ruslan Garipov <ruslanngaripov@gmail.com>